Handle required fields from schemas in item processor
Extract CSS and XPath in the SlybotIBLExtractor
Run item validation in the SlybotIBLExtractor

By running validation and selector extractors within the IBLE if no valid items
are extracted using the current sample another one can be tried. This approach
also allows for Container extractors to have nested Containers where not all
required fields need to be present in the primary Container.